Add Integration Test for MoneyTransferTypeDalService
all dependencies has to be create in DBB to create MoneyTransferType. As it has a composite primary key, we need to force, in test, the primary key of the dependencies. As we give an id before to save object, we can't further check this one before creation in DBB (moneyTransferTypeDalService.create)